chore: Adds some fluent search extensions.
Not super necessary but wraps the inverted indexes, fuzzy string searches and aliasing and sorting into more approachable linq statements.

It is worth noting that it works by constructing new inline inverted indexes and so can have performance concerns if you don't chunk the dataset early with the larger filters.

Alternatively cache the indexes at the start so as not to regenerate them inline.