chore: Adds a basic implementation of a repository
This is a pretty convoluted class but is meant as a worst case scenario catch all for a simplified ui.

If you intend to use a backing store of just an in memory list AND you have no intention of specifically providing an id for the items to use and be searched on,

This class will take a list and automatically let the index in the array act as its string key and support crud operations, even ones that don't make a lot of sense in a list context like update by id.