chore: Adds supports for queries
Now that we have repositories, this class allows us to wrap a repository in a requeste for a result.

In short, regardless of what you are searching against, where they come from and what they are...
If I give you a string and I am looking for item of type T ... return whatever amount of those items makes sense based on that string.